% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.lm.R
\name{coef.enriched_lm}
\alias{coef.enriched_lm}
\title{Function to extract model coefficients from objects of class \code{enriched_lm}}
\usage{
\method{coef}{enriched_lm}(object, model = c("mean", "full", "dispersion"), ...)
}
\arguments{
\item{object}{an object of class \code{enriched_lm}}

\item{model}{either "mean" for the estimates of the parameters in the linear predictor, or "dispersion" for the estimate of the dispersion, or "full" for all estimates}

\item{...}{currently unused}
}
\description{
Function to extract model coefficients from objects of class \code{enriched_lm}
}
