% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.link.R
\name{enrich.link-glm}
\alias{enrich.link-glm}
\title{Enrich \code{link-glm} objects}
\arguments{
\item{object}{an object of class \code{\link{family}}}

\item{with}{a character vector with the names of the components to
enrich \code{object} with. Run \code{linkglm_options(print =
TRUE)} to see the available options}

\item{...}{currently not used}
}
\value{
The object \code{object} of class
    \code{\link[=make.link]{link-glm}} with extra components. Run
    \code{linkglm_options(print = TRUE)} to see what those
    components are.
}
\description{
Enrich \code{\link[=make.link]{link-glm}} with further derivatives of
\code{linkinv} with respect to \code{eta}.
}
\details{


The \code{enrich.link-glm} method supports \code{logit},
\code{probit}, \code{cauchit}, \code{cloglog}, \code{identity},
\code{log}, \code{sqrt}, \code{1/mu^2}, \code{inverse}, as well as
the \code{\link{power}} family of links.
}
\examples{
elogit <- enrich(make.link("logit"), with = "inverse link derivatives")
str(elogit)
elogit$d2mu.deta
elogit$d3mu.deta

}

