% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\name{enrich}
\alias{enrich}
\title{Generic method for enriching objects}
\usage{
enrich(object, with, ...)
}
\arguments{
\item{object}{the object to be enriched}

\item{with}{a character vector with enrichment options for \code{object}}

\item{...}{Arguments to be passed to other methods}
}
\description{
Generic method for enriching objects
}
