% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.family.R
\name{family_options}
\alias{family_options}
\title{Available options for the enrichment of \code{family} objects}
\usage{
family_options(what, print = missing(what))
}
\arguments{
\item{what}{a character vector listing the components that should
be present in the enriched object}

\item{print}{if TRUE then the available enrichment options
are listed}
}
\description{
Available options for the enrichment of \code{family} objects
}
\details{
A check is being made whether the requested component is
    available in the \code{available_components} specification (see
    \code{\link{options_function}}). No check is being made on
    whether the functions that produce the components exist.
}
\examples{
\dontrun{
family_options(what = "all")
family_options(print = TRUE)
}
}

