% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.betareg.R
\name{get_auxiliary_functions.betareg}
\alias{get_auxiliary_functions.betareg}
\title{Function to compute/extract auxiliary functions from objects of
class \code{betreg}/\code{enriched_betareg}}
\usage{
\method{get_auxiliary_functions}{betareg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{betareg} or\code{enriched_betareg}}

\item{...}{currently not used}
}
\description{
Function to compute/extract auxiliary functions from objects of
class \code{betreg}/\code{enriched_betareg}
}
\details{
See \code{\link{enrich.betareg}} for details.
}
