% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.glm.R
\name{get_auxiliary_functions.glm}
\alias{get_auxiliary_functions.glm}
\title{Function to compute/extract auxiliary functions from objects of
class \code{glm}/\code{enriched_glm}}
\usage{
\method{get_auxiliary_functions}{glm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glm} or\code{enriched_glm}}

\item{...}{currently not used}
}
\description{
Function to compute/extract auxiliary functions from objects of
class \code{glm}/\code{enriched_glm}
}
\details{
See \code{\link{enrich.glm}} for details.
}
