% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.lm.R
\name{get_auxiliary_functions.lm}
\alias{get_auxiliary_functions.lm}
\title{Function to compute/extract auxiliary functions from objects of
class \code{lm}/\code{enriched_lm}}
\usage{
\method{get_auxiliary_functions}{lm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{lm} or\code{enriched_lm}}

\item{...}{currently not used}
}
\description{
Function to compute/extract auxiliary functions from objects of
class \code{lm}/\code{enriched_lm}
}
\details{
See \code{\link{enrich.lm}} for details.
}
