% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\name{get_auxiliary_functions}
\alias{get_auxiliary_functions}
\title{Generic method for extracting or computing auxiliary functions for
objects}
\usage{
get_auxiliary_functions(object, ...)
}
\arguments{
\item{object}{the object to be enriched or the enriched object}

\item{...}{currently not used}
}
\description{
Generic method for extracting or computing auxiliary functions for
objects
}

