% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.betareg.R
\name{get_bias_function.betareg}
\alias{get_bias_function.betareg}
\title{Function to compute/extract a function that returns the first term
in the expansion of the bias of the MLE for the parameters of an
object of class \code{betareg}/\code{enriched_betareg}}
\usage{
\method{get_bias_function}{betareg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{betareg} or\code{enriched_betareg}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a function that returns the first term
in the expansion of the bias of the MLE for the parameters of an
object of class \code{betareg}/\code{enriched_betareg}
}
\details{
The computed/extracted function has arguments
\describe{

\item{coefficients}{the regression coefficients at which the
first-order bias is evacuated. If missing then the maximum
likelihood estimates are used}
}
}
