% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\name{get_bias_function}
\alias{get_bias_function}
\title{Generic method for extracting or computing a function that returns
the bias for the parameters in modelling objects}
\usage{
get_bias_function(object, ...)
}
\arguments{
\item{object}{the object to be enriched or the enriched object}

\item{...}{currently not used}
}
\description{
Generic method for extracting or computing a function that returns
the bias for the parameters in modelling objects
}

