% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.glm.R
\name{get_dmodel_function.glm}
\alias{get_dmodel_function.glm}
\title{Function to compute/extract a \code{dmodel} function}
\usage{
\method{get_dmodel_function}{glm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glm} or\code{enriched_glm}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a \code{dmodel} function
}
\details{
The computed/extracted function has arguments
\describe{

\item{response}{an optional response vector or matrix at which to
compute densities. It must have the same shape as the response in
the fitted model. If missing, the fitted response is used. For
grouped binomial models, the row totals of the fitted response should
be retained}

\item{coefficients}{the regression coefficients at which the
densities are computed. If missing then the maximum likelihood
estimates are used}

\item{dispersion}{the dispersion parameter at which the densities
function is computed. If missing then the maximum likelihood
estimate is used}

\item{log}{logical; if \code{TRUE}, the logarithm of the density is
returned}

}
}
