% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.link-glm.R
\name{get_enrichment_options.link-glm}
\alias{get_enrichment_options.link-glm}
\title{Available options for the enrichment objects of class link-glm}
\usage{
\method{get_enrichment_options}{`link-glm`}(object, option, all_options = missing(option))
}
\arguments{
\item{object}{the object to be enriched}

\item{option}{a character vector listing the options for enriching
the object}

\item{all_options}{if \code{TRUE} then output a data frame with the
available enrichment options, their descriptions, the names of
the components that each option results in, and the names of
the corresponding \code{compute_*} functions.}
}
\value{
an object of class \code{enrichment_options}
}
\description{
Available options for the enrichment objects of class link-glm
}
\details{
A check is being made whether the requested option is
    available. No check is being made on whether the functions that
    produce the components exist.
}
\examples{
\dontrun{
`get_enrichment_options.link-glm`(option = "all")
`get_enrichment_options.link-glm`(all_options = TRUE)
}
}
