% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.betareg.R
\name{get_pmodel_function.betareg}
\alias{get_pmodel_function.betareg}
\title{Function to compute/extract a \code{pmodel} function}
\usage{
\method{get_pmodel_function}{betareg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{betareg} or\code{enriched_betareg}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a \code{pmodel} function
}
\details{
The computed/extracted function has arguments
\describe{

\item{response}{an optional numeric response vector at which to
compute the distribution function. Values must be in \code{(0, 1)},
and the vector must have the same length as the fitted response. If
missing, the fitted response is used}

\item{coefficients}{the mean and precision regression coefficients
at which the distribution function is computed. If missing, the
maximum likelihood estimates are used}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities
are \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}}

\item{log.p}{logical; if \code{TRUE}, logarithmic probabilities are
returned}

}
}
