% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.glm.R
\name{get_pmodel_function.glm}
\alias{get_pmodel_function.glm}
\title{Function to compute/extract a \code{pmodel} function}
\usage{
\method{get_pmodel_function}{glm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glm} or\code{enriched_glm}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a \code{pmodel} function
}
\details{
The computed/extracted function has arguments
\describe{

\item{response}{an optional response vector or matrix at which to
compute the distribution function. It must have the same shape as
the response in the fitted model. If missing, the fitted response
is used}

\item{coefficients}{the regression coefficients at which the
distribution function are computed. If missing then the maximum
likelihood estimates are used}

\item{dispersion}{the dispersion parameter at which the
distribution function is computed. If missing then the maximum
likelihood estimate is used}

\item{log.p}{logical; if \code{TRUE}, the logarithm of the
distribution function is returned}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities
are P[X <= x] otherwise, P[X > x]}
}
}
