% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.betareg.R
\name{get_qmodel_function.betareg}
\alias{get_qmodel_function.betareg}
\title{Function to compute/extract a \code{qmodel} function}
\usage{
\method{get_qmodel_function}{betareg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{betareg} or\code{enriched_betareg}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a \code{qmodel} function
}
\details{
The computed/extracted function has arguments
\describe{

\item{p}{a vector with one probability for each observation in the
fitted model}

\item{coefficients}{the mean and precision regression coefficients
at which the quantiles are computed. If missing, the maximum
likelihood estimates are used}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities
are \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}}

\item{log.p}{logical; if \code{TRUE}, logarithmic probabilities are
used}

}
}
