% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.glm.R
\name{get_score_function.glm}
\alias{get_score_function.glm}
\title{Function to compute/extract a function that returns the scores
(derivatives of the log-likelihood) for an object of class
\code{glm}/\code{enriched_glm}}
\usage{
\method{get_score_function}{glm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glm} or\code{enriched_glm}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a function that returns the scores
(derivatives of the log-likelihood) for an object of class
\code{glm}/\code{enriched_glm}
}
\details{
The computed/extracted function has arguments
\describe{

\item{coefficients}{the regression coefficients at which the scores
are computed. If missing then the maximum likelihood estimates are
used}

\item{dispersion}{the dispersion parameter at which the score
function is evaluated. If missing then the maximum likelihood
estimate is used}

\item{response}{an optional response vector or matrix. It must have
the same shape as the response in the fitted model. If missing, the
fitted response is used. For grouped binomial models, the row totals
of the fitted response should be retained}

}
}
