% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\name{get_simulate_function}
\alias{get_simulate_function}
\title{Generic method for extracting or computing a simulate function for
modelling objects}
\usage{
get_simulate_function(object, ...)
}
\arguments{
\item{object}{the object to be enriched or the enriched object}

\item{...}{currently not used}
}
\description{
Generic method for extracting or computing a simulate function for
modelling objects
}

