% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.link.R
\name{linkglm_options}
\alias{linkglm_options}
\title{Available options for the enrichment of \code{\link[=make.link]{link-glm}} objects}
\usage{
linkglm_options(what, print = FALSE)
}
\arguments{
\item{what}{a character vector listing the components that should
be present in the enriched object}

\item{print}{if TRUE then the available enrichment options
are listed}
}
\description{
Available options for the enrichment of \code{\link[=make.link]{link-glm}} objects
}
\details{
A check is being made whether the requested component is
    available in the \code{available_components} specification (see
    \code{\link{options_function}}). No check is being made on
    whether the functions that produce the components exist.
}
\examples{
\dontrun{
linkglm_options(what = "all")
linkglm_options(print = TRUE)
}
}

