% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_function.R
\name{options_function}
\alias{options_function}
\title{Return a template for contruction a "_options" function}
\usage{
options_function(class)
}
\arguments{
\item{class}{the class of the object for which options are to be listed}
}
\value{
A function.
}
\description{
Return a template for contruction a "_options" function
}
\details{
The function prints a new function that can be used to
    define and check for the available options for enrichment of
    the object with class \code{class}. The user is expected to
    populate the \code{available_options} vector and the
    \code{*_with} list. The latter may simply be
    \code{as.list(available_options)} if no groups of options are
    needed.
}
\examples{
\dontrun{
## family_options has been written via
options_function("family")
}
}

