% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.enrichment_options}
\alias{print.enrichment_options}
\title{Print \code{enrichment_options} objects}
\usage{
\method{print}{enrichment_options}(x, ...)
}
\description{
Print \code{enrichment_options} objects
}

