% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.glm.R
\name{coef.enriched_glm}
\alias{coef.enriched_glm}
\title{Function to extract model coefficients from objects of class \code{enriched_glm}}
\usage{
\method{coef}{enriched_glm}(object, model = c("mean", "full", "dispersion"),
  ...)
}
\arguments{
\item{object}{an object of class \code{enriched_glm}}

\item{model}{either "mean" for the estimates of the parameters in the linear predictor, or "dispersion" for the estimate of the dispersion, or "full" for all estimates}

\item{...}{currently unused}
}
\description{
Function to extract model coefficients from objects of class \code{enriched_glm}
}
