% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.family.R
\name{enrich.family}
\alias{enrich.family}
\title{Enrich \code{family} objects}
\usage{
\method{enrich}{family}(object, with = "all", ...)
}
\arguments{
\item{object}{an object of class \code{\link{family}}}

\item{with}{a character vector with the names of the components to
enrich \code{object} with. Run \code{family_options(print =
TRUE)} to see the available options}

\item{...}{currently not used}
}
\value{
The object \code{object} of class \code{\link{family}} with
    extra components. Run \code{family_options(print = TRUE)} to
    see what those components are.
}
\description{
Enrich \code{family} objects with family- and link-specfic mathematical
functions
}
\details{
\code{\link{family}} objects specify the details of the
    models used by functions such as \code{\link{glm}}. The
    families implemented in the stats package include
    \code{\link{binomial}}, \code{\link{poisson}},
    \code{\link{Gamma}}, \code{\link{inverse.gaussian}},
    \code{\link{gaussian}}, \code{\link{quasi}},
    \code{\link{quasibinomial}}, \code{\link{quasipoisson}}. All
    but the quasi families attempt to specify certain
    characteristics of distributions from the exponential family,
    which have probability mass or density function of the form
\deqn{f(y; \theta) = \exp\left\{\frac{y\theta - b(\theta) + c_1(y)}{phi/m} - a(-m/\phi) + c_2(y)\right\}}{}
where \eqn{m} is an observation weight.

These families correspond to models with \eqn{\mu = E(Y; \theta) =
b'(\theta)} and \eqn{Var(Y; \theta) = \phi V(\mu)/m} (see
\code{\link{family}} where \code{variance} corresponds to
\eqn{V(\mu)} above).

Then the deviance residuals (see \code{dev.resids}) in
\code{\link{family}}) are
\deqn{-2\left\{y c_1'(\mu) - y c_1'(y) - b(c'(\mu)) + b(c'(y))\right\}}
with the second and third derivatives of
the inverse link function with respect to \code{eta}
(\code{d2mu.eta} and \code{d3mu.eta}; the first is \code{mu.eta}),
the first and second derivative of the variance function with
respect to \code{mu} (\code{d1variance}, \code{d2variance}), and
the first three derivatives of
}
\examples{
## An example from ?glm to illustrate that things still work with
## enriched families
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
print(d.AD <- data.frame(treatment, outcome, counts))
glm.D93 <- glm(counts ~ outcome + treatment, family = enrich(poisson()))
anova(glm.D93)
summary(glm.D93)
}

