% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.glm.R
\name{enrich.glm}
\alias{enrich.glm}
\title{Enrich objects of class glm}
\usage{
\method{enrich}{glm}(object, with = "all", ...)
}
\arguments{
\item{object}{an object of class glm}

\item{with}{a character vector with the names of the components to
enrich \code{object} with}

\item{...}{extra arguments to be passed to the
\code{compute_*} functions}
}
\value{
The object \code{object} of class glm with extra
    components. \code{get_enrichment_options.glm()} returns
    the components and their descriptions.
}
\description{
Enrich of class glm with
*BRIEFLY EXPLAIN WITH WHAT*
}
\details{
*ADD DETAILS IF NECESSARY*
}
\examples{
## *ADD AN EXAMPLE*
}

