% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\docType{package}
\name{enrichwith}
\alias{enrichwith}
\alias{enrichwith-package}
\title{Methods to enrich various R objects with extra components}
\description{
The \pkg{enrichwith} package provides the \code{enrich} method to
enrich various core objects with extra, relevant components. The
current version has methods for enriching objects of class
\code{family} and \code{link-glm}. The resulting objects preserve
their class, so the methods associated to them still apply.
}
\details{
Depending on the object, enriching it can be a tedious task. The
\pkg{enrichwith} package streamlines the task into 3 simple steps:

\itemize{

\item Use \code{\link{create_enrichwith_skeleton}} to produce an
enrichwith template. This template includes all necessary functions
to carry out the enrichment.

\item Write the appropriate code for the \code{compute_*} methods

\item Finalise the documentation and/or include more examples
}
}

