% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\docType{package}
\name{enrichwith}
\alias{enrichwith}
\alias{enrichwith-package}
\title{Methods to enrich various objects with extra, relevant components}
\description{
The enrichwith package provides the "enrich" method to enrich
various core objects with extra, relevant components. The current
version has methods for enriching objects of class "family" and
"link-glm". The resulting objects preserve their class, so the
methods associated to them still apply.
}

