% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\docType{package}
\name{enrichwith}
\alias{enrichwith}
\alias{enrichwith-package}
\title{Methods to enrich various R objects with extra components}
\description{
The \pkg{enrichwith} package provides the \code{enrich} method to
enrich various objects with extra, relevant components. The
resulting objects preserve the class of the orginal objects. In
this way all methods associated to the class still apply.
}
\details{
Depending on the object, enriching it can be a tedious task. The
\pkg{enrichwith} package is an attempt to streamline the task into
3 simple steps:

\enumerate{

\item Use \code{\link{create_enrichwith_skeleton}} to produce an
enrichwith template.

\item 2. Add the specific code that calculates the components by
editing the `compute_*` functions.

\item Finalise the documentation and/or include more examples.
}

The first step results in a template that includes all necessary
functions to carry out the enrichment. The second step is where the
user edits that template and adds code to calculate the components
that the object will be enriched with. Specifically, each
\code{compute_*} function takes as input the object to be enriched
and returns the corresponding new component to be added to the
object.

Everything else (for example, mapping between the enrichment
options and the components that the enriched object will have,
checks that an enrichment option exists, listing enrichment
options, enriching the object, and so on) is taken care of by the
methods in **enrichwith**.
}

