% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-package.R
\docType{package}
\name{enrichwith}
\alias{enrichwith}
\alias{enrichwith-package}
\title{Methods to enrich various R objects with extra components}
\description{
The \pkg{enrichwith} package provides the \code{enrich} method to
enrich various core objects with extra, relevant components. The
current version has methods for enriching objects of class
\code{family} and \code{link-glm}. The resulting objects preserve
their class, so the methods associated to them still apply.
}
\details{
Depending on the object, enriching it can be a tedious task. The
\pkg{enrichwith} package streamlines the task into 3 simple steps:

\itemize{

\item Run \code{\link{build_enrichwith_skeleton}}. This step will
create a file where all functions for an \code{\link{enrich}}
template are set-up

\item Write the \code{compute_*} methods

\item Complete the documentation and/or examples
}
}

