% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.betareg.R
\name{get_information_function.betareg}
\alias{get_information_function.betareg}
\title{Function to compute/extract a function that returns the information
matrix for an object of class \code{betareg}/\code{enriched_betareg}}
\usage{
\method{get_information_function}{betareg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{betareg} or\code{enriched_betareg}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a function that returns the information
matrix for an object of class \code{betareg}/\code{enriched_betareg}
}
\details{
The computed/extracted function has arguments
\describe{

\item{coefficients}{the regression coefficients at which the
information matrix is evaluated. If missing then the maximum
likelihood estimates are used}


\item{type}{should the function return the 'expected' or 'observed' information? Default is \code{expected}}

\item{QR}{Currently not used}

\item{CHOL}{If \code{TRUE}, then the Cholesky decomposition of the information matrix at the coefficients is returned}

\item{response}{an optional numeric response vector with values in
\code{(0, 1)}. It must have the same length as the response in the
fitted model. If missing, the fitted response is used}

}
}
