% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.betareg.R
\name{get_information_function.betareg}
\alias{get_information_function.betareg}
\title{Function to compute/extract a function that returns the information
matrix for an object of class \code{betareg}/\code{enriched_betareg}}
\usage{
\method{get_information_function}{betareg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{betareg} or\code{enriched_betareg}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a function that returns the information
matrix for an object of class \code{betareg}/\code{enriched_betareg}
}
\details{
The computed/extracted function has arguments
\describe{

\item{coefficients}{the regression coefficients at which the
information matrix is evaluated. If missing then the maximum
likelihood estimates are used}

\item{QR}{If \code{TRUE}, then the QR decomposition of the expected information for the coefficients is returned}

}
}
