% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.glm.R
\name{get_qmodel_function.glm}
\alias{get_qmodel_function.glm}
\title{Function to compute/extract a \code{qmodel} function}
\usage{
\method{get_qmodel_function}{glm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{glm} or\code{enriched_glm}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a \code{qmodel} function
}

