% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.betareg.R
\name{get_score_function.betareg}
\alias{get_score_function.betareg}
\title{Function to compute/extract a function that returns the scores
(derivatives of the log-likelihood) for an object of class
\code{betareg}/\code{enriched_betareg}}
\usage{
\method{get_score_function}{betareg}(object, ...)
}
\arguments{
\item{object}{an object of class \code{betareg} or\code{enriched_betareg}}

\item{...}{currently not used}
}
\description{
Function to compute/extract a function that returns the scores
(derivatives of the log-likelihood) for an object of class
\code{betareg}/\code{enriched_betareg}
}
\details{
The computed/extracted function has arguments
\describe{

\item{coefficients}{the regression coefficients at which the scores
are computed. If missing then the maximum likelihood estimates are
used}

}
}
